requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib
numpy==1.19.5
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
QtPy==1.10.0
six==1.16.0
tk==0.1.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2